requirementts.txt
adal==1.2.7
and==66.0.2
azure-cognitiveservices-speech==1.28.0
azure-common==1.1.28
azure-core==1.26.4
azure-identity==1.12.0
azure-mgmt-automation==1.0.0
azure-mgmt-compute==29.1.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.0.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.1
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2

